+++ core/src/test/scala/ru/itclover/tsp/core/fixtures/Common.scala
@@ -16,7 +16,7 @@ object Common {
   val event: Event[Int] = Event[Int](0L, 0, 0)
 
   // Dummy event processing
-  def procEvent(ev: EInt): Long = ev.row
+  def procEvent(ev: EInt): Long = ev.row.toLong
 
   // Dummy extractor
   implicit val extractor: TsIdxExtractor[EInt] = new TsIdxExtractor(procEvent)
